Add support for Shibboleth Authentication
Allows Shibboleth OmniAuth users to login just like LDAP users.

update GemFile to load omniauth-shibboleth version 1.0.7

This is a rebase of pull request #1491 to simplify assessing its impact.

I cherry picked the original commits on top of master except for those
that adjusted the tests and the two related to images and then did an
interactive rebase to form this commit.